+++ factory-hooks/post-settings-php/mysql56-driver.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types = 1);
+
+/**
+ * Ensure MySQL 5.6 driver is configured.
+ */
+
+$databases['default']['default']['namespace'] = 'Drupal\\Driver\\Database\\mysql';